[REF] point_of_sale: move report sale details in correct file
The repport sale details has been defined in pos order file, so we are
now moving it in a better place
