+++ addons/point_of_sale/report/__init__.py
@@ -3,3 +3,4 @@
 
 from . import pos_invoice
 from . import pos_order_report
+from . import report_saledetails
